Fixing issues

+++ python/ml4ir/applications/ranking/model/losses/pointwise_losses.py
@@ -43,7 +43,7 @@ class SigmoidCrossEntropy(PointwiseLossBase):
             to the loss
         """
 
-        mask = inputs[FeatureTypeKey.MASK]
+        mask = tf.cast(inputs[FeatureTypeKey.MASK], y_pred.dtype)
 
         y_true = tf.gather_nd(y_true, tf.where(tf.equal(mask, tf.constant(1.0))))
         y_pred = tf.gather_nd(y_pred, tf.where(tf.equal(mask, tf.constant(1.0))))
